+++ src/notification.ts
@@ -124,7 +124,12 @@ export class BaseNotification {
     }
 
     const response = await this.fetch(
-      BaseNotification.getSolidWellKnownUrl(this.host)
+      BaseNotification.getSolidWellKnownUrl(this.host),
+      {
+        headers: {
+          Accept: "application/ld+json",
+        },
+      }
     );
 
     if (response.status !== 200) {